fix: separate mint card into smaller pieces
